move the tx card logic into the reusable component
